Using single entrypoint for dev-server and prod
