+++ packages/signed-api/src/index.ts
@@ -1,7 +1,11 @@
+import z from 'zod';
+
 import { fetchAndCacheConfig } from './config/config';
 import { logger } from './logger';
 import { DEFAULT_PORT, startServer } from './server';
 import { initializeVerifierPool } from './signed-data-verifier-pool';
+
+const portSchema = z.coerce.number().int().positive();
 
 const main = async () => {
   const config = await fetchAndCacheConfig();
@@ -10,7 +14,21 @@ const main = async () => {
   const pool = initializeVerifierPool();
   logger.info('Initialized verifier pool.', { maxWorkers: pool.maxWorkers, workerType: pool.workerType });
 
-  startServer(config, DEFAULT_PORT);
+  const parsedPort = portSchema.safeParse(process.env.SERVER_PORT);
+  let port: number;
+  if (parsedPort.success) {
+    port = parsedPort.data;
+    logger.debug('Using SERVER_PORT environment variable as port number.', {
+      port,
+    });
+  } else {
+    port = DEFAULT_PORT;
+    logger.debug('SERVER_PORT environment variable not set or invalid. Using default port.', {
+      port,
+    });
+  }
+
+  startServer(config, port);
 };
 
 void main();
